Fix oauth2 dependency to 1.4.7
Currently, oauth2 1.4.8 includes a change that causes aptible-cli to break when attempting authorization: https://github.com/oauth-xx/oauth2/issues/572
The underlying cause is a hash not being converted to a string, which Faraday's default adapter provides. This will need to be fixed in the oauth2 gem to properly address the issue.